+++ tests/Firefly.CrossPlatformZip.Tests.Unit/UnzipTests.cs
@@ -6,7 +6,6 @@ namespace Firefly.CrossPlatformZip.Tests.Unit
     using System;
     using System.IO;
     using System.Linq;
-    using System.Runtime.InteropServices;
     using System.Security.Cryptography;
 
     using FluentAssertions;