+++ tests/Firefly.CrossPlatformZip.Tests.Unit/ZipTests.cs
@@ -92,7 +92,7 @@ namespace Firefly.CrossPlatformZip.Tests.Unit
         }
 
         /// <summary>
-        /// Given a single file and using <see cref="Zipper.Zip(string,string, int)"/> method then it is added at the root of the central directory.
+        /// Given a single file and using <see cref="Zipper.Zip(CrossPlatformZipSettings)"/> method then it is added at the root of the central directory.
         /// </summary>
         [TestMethod]
         public void GivenASingleFile_AndUsingZipMethod_ThenItIsAddedAtTheRootOfTheCentralDirectory()
@@ -118,7 +118,7 @@ namespace Firefly.CrossPlatformZip.Tests.Unit
         }
 
         /// <summary>
-        /// Given a single file and using <see cref="Zipper.ZipSingleFile(string,string,string,int)"/> method with the alternate name argument = <c>null</c>
+        /// Given a single file and using <see cref="Zipper.ZipSingleFile(CrossPlatformZipSettings)"/> method with the alternate name argument = <c>null</c>
         /// then it is added at the root of the central directory.
         /// </summary>
         [TestMethod]
@@ -145,7 +145,7 @@ namespace Firefly.CrossPlatformZip.Tests.Unit
         }
 
         /// <summary>
-        /// Given a single file and using <see cref="Zipper.ZipSingleFile(string,string,string,int)"/> method with the alternate name argument not <c>null</c>
+        /// Given a single file and using <see cref="Zipper.ZipSingleFile(CrossPlatformZipSettings)"/> method with the alternate name argument not <c>null</c>
         /// then it is added at the root of the central directory with the alternate filename.
         /// </summary>
         [TestMethod]