+++ RELEASENOTES.md
@@ -1,3 +1,8 @@
+## 0.5.0
+
+* Refactor `ExternalAttributes` to be `Traits` since this is really what these classes represent.
+* Add a trait to detect file name/directory name duplication when creating a zip for Windows from a case sensitive file system. Throw an exception if detected.
+
 ## 0.4.1
 
 When adding files from a unix filesystem to a windows zip, store the MS-DOS attributes as Archive. 
@@ -9,7 +14,7 @@ Reworked public interface to take a settings class as there were too many argume
 
 ## 0.3.0
 
-Basic implementataion of Unix Extra Fields in central directory entries of zips targeting unix-like operating systems.
+Basic implementation of Unix Extra Fields in central directory entries of zips targeting unix-like operating systems.
 
 * Unix Extra type 3 (tag Id 0x7875) - Stores uid/gid. Presently this is set to 0/0 until netcore3.0 which suppots POSIX properly
 * Extended Timestamp (tag ID 0x5455) - Timestamps are stored but not yet restored by this unzip implementation.